refactor: extract mesh tests to separate files
Move ~4,900 lines of tests out of mesh/mod.rs into:
- mesh/tests.rs (4,823 lines) — main test module
- mesh/public_identity_tests.rs (70 lines) — public identity lifecycle tests

mod.rs goes from 10,351 to 5,460 lines. Pure mechanical move,
no logic changes. All 808 tests pass.
